Was able to cut out a LOT of duplication among the 4 Mission fragment classes
